Make required changes for this repo to be tested by core jenkins CI instance
We are migrating all jenkins jobs from our blue instance to our core jenkins CI
instance.  This requires that we..

1. Run on a specified jenkins agent for the owning team
2. Use the `sshagent` wrapper to get access to git credentials

After this change is reviewed and merged, we will create a new jenkins pipeline
job for it on the core jenkins CI instance, currently hosted at
https://tools-jenkins-us-west-2.forge.auth0.net .  When that happens, it will
automatically run a build against the master branch. This will cause the publish
step to execute and we will need to confirm that it does not break anything
downstream and that it runs correctly.